Create _bonusMain.c
Send it to _bonusMain() from _bonjsGetCommandsNum. I hope it will fix freeing of fd array, cause I found problem in that function: in the end of program I freeing fd by commands_num Las a counter. But it was sometimes smaller than in the start of allocation memory for fd array.